fix: capture multipart cleanup and settle reader failures
